pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "negotiator"
version = "0.1.0"
description = "Reinforcement learning multi-agent negotiation simulator built on NegMAS"
readme = "README.md"
authors = [{name = "Negotiation Lab"}]
requires-python = ">=3.10"
dependencies = [
    "negmas>=0.11.3",
    "numpy>=1.26",
    "torch>=2.2",
    "gymnasium>=0.29",
    "pyyaml>=6.0",
    "tqdm>=4.66"
]

[project.optional-dependencies]
development = ["pytest>=8.0", "black>=24.0", "ruff>=0.6"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
